# specs/openapi.yaml
openapi: 3.1.0
info:
  title: bracket-creator API
  description: |
    Web API for generating Kendo tournament brackets.

    This spec covers two servers:
    - **Bracket Generator** (`serve` command): Stateless, form-based bracket generation.
    - **Mobile/Admin API** (`mobile-app` command): Stateful REST API for live tournament management.

    The Mobile/Admin API routes under `/api/` (excluding `/api/viewer/`) require the
    `X-Tournament-Password` header for authentication.
  version: 2.1.0
servers:
  - url: http://localhost:8080
    description: Bracket Generator (serve command)
  - url: http://localhost:8081
    description: Mobile/Admin API (mobile-app command)
...
paths:
...
  /api/tournament:
    get:
      tags: [tournament]
      summary: Get tournament configuration
      security:
        - TournamentPassword: []
      responses:
        '200':
          description: Tournament configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [tournament]
      summary: Create tournament configuration
      description: |
        Creates a new tournament. Allowed without the `X-Tournament-Password` header when no
        tournament has been configured yet (bootstrap). Once a tournament with a password exists,
        auth is required.

        Validation: `name` is trimmed and must be non-empty after trim (`tournament name is
        required`). `password` must be non-empty (`tournament password is required`) — empty
        Password on disk would let AuthMiddleware's `password != t.Password` check vacuously
        pass for any client sending an empty header. `venue` and `date` are also trimmed but
        may be empty.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '201':
          description: Tournament created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequest'
    put:
      tags: [tournament]
      summary: Update tournament configuration
      description: |
        Updates tournament config. Also allowed without auth during bootstrap (same
        condition as POST): `AuthMiddleware` skips the password check only when
        `tournament.md` is missing OR when the on-disk record is the literal default
        placeholder (`Name == "New Tournament" && Password == ""`). A real-named
        record with an empty password is FAIL-CLOSED — rejected with 403
        (`tournament misconfigured: password is not set`) BEFORE this handler runs;
        operator must repair `tournament.md` out-of-band to recover. See
        `internal/mobileapp/middleware.go` AuthMiddleware for the exact branches.

        Validation matches POST for `name` (rejected if empty after trim).

        **Password preserve semantics**: when the request body omits `password` or sends an
        empty string, the server preserves the currently-stored password. This matches the
        admin UI's `password: pass || undefined` pattern ("leave blank to keep current").
        If after the preserve step the password is STILL empty (operator manually edited
        `tournament.md`, or a fresh PUT against a never-initialized tournament), the
        request is rejected with 400 (`tournament password is required`) to force the
        admin to set a password rather than silently allow the empty-stored-password
        vacuous-pass scenario.
      security:
        - TournamentPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '200':
          description: Tournament updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    TournamentPassword:
      type: apiKey
      in: header
      name: X-Tournament-Password
      description: |
        Tournament admin password. Required for all `/api/` routes except
        `/api/viewer/`, `/api/events`, and the bootstrap PUT/POST
        `/api/tournament` while no tournament has been configured yet.

        Header value must exactly equal the stored tournament password
        (exact-string match; no trimming applied to either side).

        Defense-in-depth: if the stored tournament password is somehow
        empty (e.g. operator manually edited `tournament.md` to land
        an empty `password:` field), AuthMiddleware returns 403
        (`tournament misconfigured: password is not set`) for ALL
        requests rather than letting the `header != stored` check
        vacuously pass for empty header.

        Recovery from this state requires **out-of-band repair**: the
        bootstrap PUT/POST `/api/tournament` exception only applies
        when the loaded record is `nil` or matches the literal
        `Name == "New Tournament" && Password == ""` default — a
        real-named record with empty Password is locked out by the 403
        fail-closed branch. Operator must stop the server, edit
        `tournament.md` to either delete it (forces fresh bootstrap)
        or set a non-empty `password:` field in the YAML front-matter,
        then restart.
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Tournament:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          description: Tournament name. Trimmed server-side; rejected with 400 if empty after trim or longer than 200 characters.
        date:
          type: string
          maxLength: 10
          description: >
            Event date in DD-MM-YYYY format. Trimmed server-side; rejected
            with 400 if non-empty and not a syntactically valid day in
            that format (Feb 31 etc. → 400) OR if the year falls outside
            1900–2100. The year-range matches the frontend validator
            (web-mobile/js/admin_helpers.jsx MIN_YEAR/MAX_YEAR) so a
            value the API accepts is also one the admin UI can edit. The
            web frontend converts the HTML `<input type="date">` ISO
            output to DMY at the input boundary; direct API clients
            send DMY directly.
        venue:
          type: string
          maxLength: 200
          description: Venue name. Trimmed server-side.
        courts:
          type: array
          items:
            type: string
          description: >
            Court labels (e.g. ["A", "B"]). Each label is a single
            character; 26-court cap (A–Z). Duplicate labels are rejected
            with 400 — the frontend uses labels as identity keys (per-court
            rendering, schedule `byCourt` bucketing, filter values) and
            duplicates would collapse data structures.
        password:
          type: string
          maxLength: 256
          description: >
            Admin password. Required (non-empty) on POST /tournament — an
            empty Password on disk would let AuthMiddleware's
            `password != t.Password` check vacuously pass for any client
            sending an empty `X-Tournament-Password` header.

            On PUT /tournament: omitting the field or sending an empty
            string preserves the currently-stored password ("leave blank
            to keep current"). Sending a non-empty string updates it.

            Omitted from viewer responses (`/api/viewer/tournament`).
        openingBlock:
          type: string
          maxLength: 16
          description: Opening ceremony block label/duration.
        lunchBlock:
          type: string
          maxLength: 16
          description: Lunch break block label/duration.
        closingBlock:
          type: string
          maxLength: 16
          description: Closing ceremony block label/duration.
        clockToElapsedMultiplier:
          type: number
          description: Multiplier converting on-clock match minutes to wall-clock elapsed minutes.
        slowestCourtBufferPct:
          type: integer
          description: Buffer percentage applied to the slowest court when estimating overall duration.
...